Enable async datasource refresh for all catalogs (#67)
This PR enables openhexa to delay sync() for all catalogs, and do the work in a separate process (called sync_datasources_worker). This change of behavior is activated with a new settings, DATASOURCE_ASYNC_REFRESH, which is False by default (no change of behavior, no change of deploy settings needed). It use the django-postgres-queue because it doesnt need a lot of new infrastructure (like celery or something else) and is suffisant for the immediate future.