chore: Keep comm with JupyterHub API internal to the Docker network (#599)
When working with Docker (locally), the backend has to communicate with the
JupyterHub API. Previously, it used the host URL as JupyterHub is exposed
on its port 8001. It is not necessary to go through the host. The communication
can be done to the Docker network dedicated to OpenHexa services.

Issue BLSQ/openhexa-team#617

Co-authored-by: Christophe Philemotte <1059797+toch@users.noreply.github.com>