chore: Make the Docker image "rootless" (#799)
The previous image set root as the default user. This means that a process
has the root rights and can create files with this id in a mounted volume from
the host. This is a common attack vector.

With the present change, the container will run under a normal user `openhexa`
without any root permission.

It's also possible to set the UID and GID of the containerized process and
session. The pipeline runner requires the GID that corresponds to the host
Docker group.

Noticed that doesn't make Docker rootless. This requires to prepare the Docker
engine to run in a normal userspace. This goes beyond the present change.

Co-authored-by: Christophe Philemotte <1059797+toch@users.noreply.github.com>