Fix: react update + Fix docker-compose (#991)
* fix(docker): update context and volume paths for pipelines services in docker-compose.yaml

* chore: downgrade React to 18

It is not supported by react-popper & react-table

Both libraries are not maintained anymore

* chore: downgrade @types/react to 18 and add @types/prop-types

Updated the type definitions for React to version 18.0.0 due to compatibility issues, and added @types/prop-types to the dependencies. Additionally, updated the Dialog component to specify the type for the ref, and adjusted the Tabs component to ensure proper type casting for child elements. Modified the cron expression validation method to use the correct parsing function.