Add application_name to DB connections from pipelines
For connections to the workspace DB, we add the `application_name`
parameter to provide insights on which pipeline runs which query.
This will allow us to see which pipelines are responsible for slow
queries for example.
This works for both Python as well as notebook pipelines.
It does not affect the regular connection string or the one in the
notebook.