feat: Add application_name parameter to Postgres connections from Pipelines (HEXA-1419) (#1460)
* remove truncate on the DB credentials

If I click "see" I want to see it all.

* Add application_name to DB connections from pipelines

For connections to the workspace DB, we add the `application_name`
parameter to provide insights on which pipeline runs which query.
This will allow us to see which pipelines are responsible for slow
queries for example.
This works for both Python as well as notebook pipelines.
It does not affect the regular connection string or the one in the
notebook.

* Add some tests

* Fix: ruff

* revert: Don't add application_name to external DB connections

This code related to the "Connections", and for now we just want to
track the connections to the workspace DB. Let's remove this unnecessary
complexity for now.

* feat: Add the application_name also for notebooks connections

* fix: Add workspace slug to application_name